+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+* Added some Anvil chances in Other Experience feature
+  * Added "Remove Too Expensive". As the name suggests, removes the 39 levels cap from the anvil
+  * Added "Remove rename cost". Renaming items is now always free
+
 ## 2.12.1
 * Added `Eating Time Minimum` config option, defaulting to 24 (from 32)
 * Power enchantment power can now be configured